refactor(evm): based on ai review, add changeset
