+++ .changeset/rotten-suits-strive.md
@@ -1,0 +1,5 @@
+---
+"@venusprotocol/evm": minor
+---
+
+Vault page revamp
